requirements.txt
pytest
pytest-dotenv
python-dotenv
fastapi
uvicorn
jwt
pandas
sqlalchemy
pydantic
psycopg2-binary
sqlalchemy-cockroachdb
opentelemetry-exporter-otlp
opentelemetry-instrumentation-psycopg2
opentelemetry-instrumentation-logging
opentelemetry-instrumentation-fastapi